fix(tests): harden ringtail+protocol+harness against parallel-run flakes
Three independent flakes turned `go test ./...` red under default
parallelism. Root causes and fixes:

1. ringtail upstream race: github.com/luxfi/ringtail v0.2.0 stores its
   precomputed-randomness buffer in package globals
   (PrecomputedRandomness, RandomnessIndex), so two goroutines calling
   threshold.GenerateKeys concurrently slice past capacity and panic with
   "slice bounds out of range". Serialize the wrapper call site in
   protocols/ringtail/keygen/round1.go behind a package-level Mutex
   (UpstreamMu) until upstream is goroutine-safe.

2. ringtail config-test type mismatch: assert.Greater(t, params.Q, 0)
   compared a uint64 against an untyped int literal, which testify
   refuses ("Elements should be the same type"). Use uint64(0).

3. pkg/protocol mockRound concurrent map writes: handler_test.go's
   mock StoreMessage / StoreBroadcastMessage wrote to its messages /
   broadcasts maps without locking, but the production handler dispatches
   message processing across worker goroutines. Add sync.Mutex on
   mockRound and lock both store paths.

Also raise the default test harness timeout from 180s to 5m. Frost 5-of-7
keygen finishes in <1s wall-clock alone, but when go test launches every
package in parallel (cmp + frost + protocols + paillier all CPU-bound at
once), the harness context expires before the protocol does. The
per-handler ProtocolTimeout was already 5m; align the harness wrapper.

Tests: 58 packages OK, 0 FAIL across 3 consecutive `go test ./...` runs.

+++ protocols/ringtail/ringtail_test.go
@@ -276,7 +276,7 @@ func TestConfigValidation(t *testing.T) {
 
 				params := cfg.GetParameters()
 				assert.Greater(t, params.N, 0)
-				assert.Greater(t, params.Q, 0)
+				assert.Greater(t, params.Q, uint64(0))
 				assert.Greater(t, params.Sigma, 0.0)
 			}
 		})
@@ -297,7 +297,7 @@ func TestSecurityLevels(t *testing.T) {
 
 		params := cfg.GetParameters()
 		assert.Greater(t, params.N, 0)
-		assert.Greater(t, params.Q, 0)
+		assert.Greater(t, params.Q, uint64(0))
 		assert.Greater(t, params.Sigma, 0.0)
 
 		switch level {
